Organize the code better and move the constants in a separate package.
