--- st2actions/st2actions/constants.py
@@ -1,1 +1,0 @@
-LIBS_DIR = 'lib'
