+++ st2actions/st2actions/container/base.py
@@ -4,8 +4,8 @@ import traceback
 
 from st2common import log as logging
 from st2common.exceptions.actionrunner import ActionRunnerCreateError
-from st2common.models.api.constants import (ACTIONEXEC_STATUS_SUCCEEDED,
-                                            ACTIONEXEC_STATUS_FAILED)
+from st2common.constants.action import (ACTIONEXEC_STATUS_SUCCEEDED,
+                                        ACTIONEXEC_STATUS_FAILED)
 from st2common.services import access
 from st2common.util.action_db import (get_action_by_dict, get_runnertype_by_name)
 from st2common.util.action_db import (update_actionexecution_status, get_actionexec_by_id)
