+++ st2actions/st2actions/runners/pythonrunner.py
@@ -12,7 +12,7 @@ from eventlet import greenio
 from multiprocessing import Process
 from st2actions.runners import ActionRunner
 from st2common import log as logging
-from st2common.models.api.constants import ACTIONEXEC_STATUS_SUCCEEDED, ACTIONEXEC_STATUS_FAILED
+from st2common.constants.action import ACTIONEXEC_STATUS_SUCCEEDED, ACTIONEXEC_STATUS_FAILED
 from st2common.util import loader as action_loader
 from st2common.util.config_parser import ContentPackConfigParser
