+++ st2actions/st2actions/worker.py
@@ -5,7 +5,7 @@ from oslo.config import cfg
 from st2actions.container.base import RunnerContainer
 from st2common import log as logging
 from st2common.exceptions.db import StackStormDBObjectNotFoundError
-from st2common.models.api.constants import (ACTIONEXEC_STATUS_RUNNING, ACTIONEXEC_STATUS_FAILED)
+from st2common.constants.action import (ACTIONEXEC_STATUS_RUNNING, ACTIONEXEC_STATUS_FAILED)
 from st2common.exceptions.actionrunner import ActionRunnerException
 from st2common.transport import actionexecution, publishers
 from st2common.util.action_db import (get_actionexec_by_id, update_actionexecution_status)
