+++ st2actions/tests/test_actionchain.py
@@ -8,9 +8,9 @@ from unittest2 import TestCase
 from st2actions.runners import actionchainrunner as acr
 from st2actions.container.service import RunnerContainerService
 from st2common.exceptions import actionrunner as runnerexceptions
-from st2common.models.api.constants import ACTIONEXEC_STATUS_RUNNING
-from st2common.models.api.constants import ACTIONEXEC_STATUS_SUCCEEDED
-from st2common.models.api.constants import ACTIONEXEC_STATUS_FAILED
+from st2common.constants.action import ACTIONEXEC_STATUS_RUNNING
+from st2common.constants.action import ACTIONEXEC_STATUS_SUCCEEDED
+from st2common.constants.action import ACTIONEXEC_STATUS_FAILED
 from st2common.services import action as action_service
 from st2common.util import action_db as action_db_util
 import st2tests.config as tests_config
