+++ st2client/st2client/commands/action.py
@@ -7,8 +7,8 @@ import time
 import six
 import sys
 
-from st2common.models.api.constants import ACTIONEXEC_STATUS_SCHEDULED
-from st2common.models.api.constants import ACTIONEXEC_STATUS_RUNNING
+from st2common.constants.action import ACTIONEXEC_STATUS_SCHEDULED
+from st2common.constants.action import ACTIONEXEC_STATUS_RUNNING
 
 from st2client import models
 from st2client.commands import resource
